+++ packages/block-library/src/cover/edit/block-controls.js
@@ -9,7 +9,7 @@ import {
 	__experimentalBlockAlignmentMatrixControl as BlockAlignmentMatrixControl,
 	__experimentalBlockFullHeightAligmentControl as FullHeightAlignmentControl,
 } from '@wordpress/block-editor';
-import { __ } from '@wordpress/i18n';
+import { __, isRTL } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
@@ -58,15 +58,27 @@ export default function CoverBlockControls( {
 		} );
 	};
 
+	// Flip value horizontally to match the physical direction indicated by
+	// AlignmentMatrixControl with the logical direction indicated by cover
+	// block in RTL languages.
+	const flipHorizontalPosition = ( ltrContentPosition ) => {
+		return isRTL()
+			? ltrContentPosition.replace( /left|right/, ( match ) =>
+					match === 'left' ? 'right' : 'left'
+			  )
+			: ltrContentPosition;
+	};
+
 	return (
 		<>
 			<BlockControls group="block">
 				<BlockAlignmentMatrixControl
 					label={ __( 'Change content position' ) }
-					value={ contentPosition }
+					value={ flipHorizontalPosition( contentPosition ) }
 					onChange={ ( nextPosition ) =>
 						setAttributes( {
-							contentPosition: nextPosition,
+							contentPosition:
+								flipHorizontalPosition( nextPosition ),
 						} )
 					}
 					isDisabled={ ! hasInnerBlocks }